requirements.txt
Authlib==1.6.1
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
cryptography==45.0.6
elastic-transport==9.1.0
elasticsearch==9.1.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
tornado==6.5.1
typing_extensions==4.14.1
urllib3==2.5.0
